# PYTHON-3824 Optimize BSON encoding of standard Python list and tuples