--- a/bson/_cbsonmodule.c
+++ b/bson/_cbsonmodule.c
@@ -859,6 +859,7 @@ static int _write_element_to_buffer(PyObject* self, buffer_t buffer,
     PyObject* new_value = NULL;
     int retval;
     PyObject* uuid_type;
+    int is_list;
     /*
      * Don't use PyObject_IsInstance for our custom types. It causes
      * problems with python sub interpreters. Our custom types should
@@ -1138,8 +1139,7 @@ static int _write_element_to_buffer(PyObject* self, buffer_t buffer,
     } else if (PyDict_Check(value)) {
         *(pymongo_buffer_get_buffer(buffer) + type_byte) = 0x03;
         return write_dict(self, buffer, value, check_keys, options, 0);
-    } else if (PyList_Check(value) || PyTuple_Check(value)) {
-        int is_list = PyList_Check(value);
+    } else if ((is_list = PyList_Check(value)) || PyTuple_Check(value)) {
         Py_ssize_t items, i;
         int start_position,
             length_location,
@@ -1188,8 +1188,9 @@ static int _write_element_to_buffer(PyObject* self, buffer_t buffer,
             } else {
                 item_value = PyTuple_GetItem(value, i);
             }
-            if (!item_value)
+            if (!item_value) {
                 return 0;
+            }
             if (!write_element_to_buffer(self, buffer, list_type_byte,
                                          item_value, check_keys, options,
                                          0, 0)) {