https://jira.mongodb.org/browse/PYTHON-3824

Manual testing shows the following results:
before:
```
python -m timeit -s 'from bson import encode;doc={"a": list((["a"], "b") for _ in range(1000))}' 'encode(doc)'
2000 loops, best of 5: 170 usec per loop

python -m timeit -s 'from bson import encode;doc={"a": list(range(1000))}' 'encode(doc)'
10000 loops, best of 5: 38.9 usec per loop
```
after:
```
python -m timeit -s 'from bson import encode;doc={"a": list((["a"], "b") for _ in range(1000))}' 'encode(doc)'
2000 loops, best of 5: 159 usec per loop

python -m timeit -s 'from bson import encode;doc={"a": list(range(1000))}' 'encode(doc)'
10000 loops, best of 5: 38.5 usec per loop
```