@sleepyStick 2023-07-13
Sorry, I had previously posted the numbers in the JIRA ticket, but I just added them here as well.