@ShaneHarvey 2023-07-13
Thanks, can you test a regular list of numbers as well:
```
python -m timeit -s 'from bson import encode;doc={"a": list(range(1000))}' 'encode(doc)'
```