@sleepyStick 2023-07-13
yes! just added that here and in the JIRA ticket